refactor(view/app): phase5-2 split debug scene by use-case
