+++ view/app/src/app_state/settings_accessors/snapshot_settings.rs
@@ -6,7 +6,7 @@ impl AppState {
     pub fn set_snapshot_work_solid_color(&mut self, color: [f32; 4]) {
         self.snapshot_shaded_color_settings.work_solid_color = color;
 
-        if self.debug_snapshot_shaded_mode {
+        if self.debug_snapshot.shaded_mode {
             let stage = self.renderer.get_stage_mut();
             if let Some(mesh_stage) = stage.as_any_mut().downcast_mut::<MeshStage>() {
                 mesh_stage.set_mesh_base_color(color);
@@ -18,7 +18,7 @@ impl AppState {
     pub fn set_snapshot_tool_wire_color(&mut self, color: [f32; 3]) {
         self.snapshot_shaded_color_settings.tool_wire_color = color;
 
-        if let Some(tool_lines_per_frame) = &mut self.debug_snapshot_tool_lines {
+        if let Some(tool_lines_per_frame) = &mut self.debug_snapshot.tool_lines {
             for frame in tool_lines_per_frame.iter_mut() {
                 for vertex in frame.iter_mut() {
                     vertex.normal = color;
@@ -26,7 +26,7 @@ impl AppState {
             }
         }
 
-        if self.debug_snapshot_shaded_mode {
+        if self.debug_snapshot.shaded_mode {
             self.sync_snapshot_visual_frame();
         }
     }
